remove PTEF_IGNORE_FILES and any exclude/ignore logic
The use case for multi-hierarchy-level ignores is very narrow - a much
more common use case would be ignoring some files *only* in one
specific directory. This could be easily achieved with char **excludes
(or -e switch, multiple times) in the runner implementation, while not
being part of the spec.

However, the problem is deeper than just the env variable.

The use case of ignoring arbitrary files is actually almost nonexistent
as there are others ways to mitigate it:

 1) utilities can be in utils/ without a 'run' to execute them

 2) a small number of executables/subdirectories can be just specified
    as ptef-runner arguments in a runner wrapper IF that wrapper itself
    is running without arguments, *not* passing the files that are to
    be excluded

 3) any bigger number of excluded files is even rarer and could be done
    with a custom find-based logic and piped into xargs ptef-runner

All these would be easier if the runner had an exclude option, however
there are many more similar cases - what if the user wants to exclude
based on a shell glob? A regular expression?
Supporting these in the .c runner would be infeasible.

Simple excludes like (1) and (2) can be easily done in a runner
wrapper, and (3) and anything beyond is probably best served by
a custom runner, possibly utilizing ptef_report / ptef_mklog.

A char **excludes argument would just pollute the API.

Signed-off-by: Jiri Jaburek <comps@nomail.dom>