add back runner ignore functionality
There *is* one important use case - ignoring non-argv0-named runners.

A fairly common use case might be to have 'build', 'run' and 'clean'
executables that traverse the same hierarchy and perform different
tasks, and have a shared runner that ignores all 3 of these and that
is itself symlinked across the hierarchy.

This means the user cannot just explicitly list all the other
non-build non-run non-clean executables in CWD, because the runner
is shared (symlinked) everywhere.

While PTEF_IGNORE_FILES would have solved this too, it was a hammer
too heavy for what it needed to do - a user might want to opt-out
of the ignores for a specific subdirectory, so it's a good idea to
*not* propagate the ignores.

Symlinking a shared runner is, however, very easy to do, so
a current-level-only ignore logic can easily be extended to be
multi-level *while* providing the flexibility of not always
propagating downwards.

Signed-off-by: Jiri Jaburek <comps@nomail.dom>