get rid of -Werror due to compiler bugs
Both gcc (on s390x) and clang (even on x86) report false positives
with -Wextra. While I can deal with these, a random user who wants
to build the source in a VM/container might have only clang and
thus fail the build and be confused.

For the sake of builds working everywhere, don't error on warnings.

Signed-off-by: Jiri Jaburek <comps@nomail.dom>