remove __thread from ptef_status_colors
This is a gcc/clang extension, thread-local variables came later
in C11 as _Thread_local, but - even then - I imagine the C ABI
support for them would be fairly wild.

Since it is fairly unlikely that a user might want different color
palettes used in different threads of the same process, just make
the variable a regular global symbol.

Signed-off-by: Jiri Jaburek <comps@nomail.dom>